\name{affy-deprecated}
\alias{loess.normalize}
\title{Deprecated functions in package \sQuote{affy}}
\description{
  These functions are provided for compatibility with older versions
  of affy only, and will be defunct at the next release.
}
\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{loess.normalize: \code{\link{normalize.loess}}}

  }
}
