\name{affybatch.example}
\docType{instance}
\alias{affybatch.example}
\title{AffyBatch instance affybatch.example}
\description{
  This is an artifical data set. It shows values subset from the
  Dilution dataset to make 100x100 chips.
}
%\usage{data(affybatch.example)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} of three samples.}
\code{\link{cdfenv.example}}}
\seealso{\code{\link[affydata]{Dilution}},\code{\link{cdfenv.example}}
}
\keyword{methods}



