\name{cdfenv.example}
\alias{cdfenv.example}
\alias{cdf.example}
\title{Example of Cdf and cdfenv}
\description{
  Example of Cdf object and its corresponding cdfenv
  (environment containing the probe locations).
}
\usage{data(cdfenv.example)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} object 'affybatch.example' containing experimental data and an
  \code{\link{environment}} 'cdf.example' containing the probe
  locations}
\source{Affymetrix CDF file for the array Hu6800}
\seealso{affybatch.example}
\keyword{datasets}
