\name{Dilution}
\docType{instance}
\alias{Dilution}
\title{AffyBatch instance Dilution}
\description{
This \code{\link[affy]{AffyBatch-class}} represents part of dilution experiment data set.
}
\usage{data(Dilution)}
\format{A \code{\link[affy]{AffyBatch-class}} containing 4 arrays.}
\source{Two sources of cRNA A (human liver tissue) and B (Central Nervous
System cell line) have been hybridized to human array (HGU95A) in a range
of proportions and dilutions. This data set is taken from
arrays hybridized to source A at 10.0 and 20 $\mu$g. We have two
replicate arrays 
for each generated cRNA. Three scanners have been used
in this study. Each array replicate was 
processed in a different scanner.
This is the identical data set that appears in the affydata package,
but is duplicated here to prevent a recursive package dependency.
For more information see Irizarry, R.A., et al. (2001) \url{http://www.biostat.jhsph.edu/~ririzarr/papers/index.html}}


\keyword{datasets}
