\name{generateExprVal.method.avgdiff}
\alias{generateExprVal.method.avgdiff}
\title{Generate an expression value from the probes informations}
\description{
  Generate an expression from the probes
}
\usage{
generateExprVal.method.avgdiff(probes)
}
\arguments{
  \item{probes}{XXX}
}
\value{
  XXX
}
\seealso{
  \code{\link{generateExprVal.method.playerout}}
}
\keyword{manip}
