\name{probeMatch-methods}
\docType{methods}
\title{Methods for accessing perfect matches and mismatches}
\alias{probeMatch-methods}
\alias{probeMatch}
\alias{pm}
\alias{pm<-}
\alias{mm}
\alias{mm<-}
\description{Methods for perfect matches and mismatches probes}
\section{Methods}{\describe{
    \item{object = AffyBatch}{All the \emph{perfect match} (pm) or
      \emph{mismatch} (mm) probes on the
      arrays the object represents are returned.}
    \item{object = ProbeSet}{The \code{pm} or \code{mm} of the object are returned}
  }
}
\keyword{methods}
