\name{probeNames-methods}
\docType{methods}
\title{Methods for accessing the Probe Names}
\alias{probeNames-methods}
\alias{probeNames}
\description{Methods for accessing Probe Names}
\section{Methods}{\describe{
\item{object = Cdf}{ An accesor function for the
      \code{name} slot.}
\item{object = PPSet.container}{ An accesor function for the
      \code{name} slot of the Cdf that goes with the container.}
\item{object = Plob}{  An accesor function for the
      \code{name} slot. }
}}
\keyword{methods}
