\name{summary}
\alias{summary}
\alias{medianpolish}
\alias{tukeybiweight}
\alias{avdiff}
\title{Probe Set Summarizing Functions}
\description{
  These were used with the function \code{express} which is no longer
  part of the package. Some are still used by the generateExprVal functions.
  But you should avoid using them directly.
}  
\seealso{\code{\link{expresso}}}

\keyword{manip}% at least one, from doc/KEYWORDS





