\name{affybatch.example}
\docType{data}
\alias{affybatch.example}
\alias{affybatch.example2}
\title{AffyBatch instance affybatch.example}
\description{
  This is an artifical data set. It contains two similar examples
  (\code{affybatch.example}, \code{affybatch.example2}) with
  values subset from the Dilution dataset to make two 100x100
  chips. Either example is suitable for exploring \code{affy}; both are
  used to demonstrate \code{\link{merge.AffyBatch}}.
}
%\usage{data(affybatch.example)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} of three samples.}
%\code{\link{cdfenv.example}}
\seealso{\code{\link[affydata]{Dilution}},\code{\link{cdfenv.example}}
}
\keyword{methods}
