\name{affybatch.example}
\docType{instance}
\alias{affybatch.example}
\title{AffyBatch instance affybatch.example}
\description{
  This data set gives a subset of the values read from a CEL file from a HU6800
  chip.
}
\usage{}%data(affybatch.example)}
\format{An \code{\link{AffyBatch}} of three samples.}
\source{Data taken from Lemon et al. [Lemon, WJ, Palatini, JT, Krahe, R, Wright, FA,
"Comparison of gene expression estimators for oligonucleotide arrays",
submitted 2001.]. The locations are given by the environmemt
\code{\link{corner.Hu6800.example}}}
\seealso{\code{\link{Dilution}},\code{\link{corner.Hu6800.example}}
}
\keyword{methods}



