\name{cdfenv.example}
\alias{cdfenv.example}
\title{Example cdfenv}
\description{
  Example cdfenv
  (environment containing the probe locations).
}
\usage{data(cdfenv.example)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} object
  \code{affybatch.example} containing experimental data and an
  \code{\link{environment}} \code{cdfenv.example} containing the probe
  locations}
\source{Affymetrix CDF file for the array Hu6800}
\seealso{affybatch.example}
\keyword{datasets}
