\name{cdfenv.example}
\alias{cdfenv.example}
\title{Example cdfenv}
\description{
  Example cdfenv (environment containing the probe locations).
}
\usage{data(cdfenv.example)}
\format{ An
  \code{\link{environment}} \code{cdfenv.example} containing the probe
  locations}
\source{Affymetrix CDF file for the array Hu6800}
\keyword{datasets}
