\name{probeNames-methods}
\docType{methods}
\title{Methods for accessing the Probe Names}
\alias{probeNames-methods}
\alias{probeNames}
\alias{probeNames<-}
%\alias{probeName}
%\alias{probeName<-}
\description{Methods for accessing Probe Names}
\section{Methods}{\describe{
\item{object = Cdf}{ An accesor function for the
      \code{name} slot.}
\item{object = probeNames}{Returns the probe names associated with the
  rownames of the intensity matrices one gets with the \code{pm} and
  \code{mm} methods.}
}}
\keyword{methods}
